# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Core App Rest",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/categories": {
            "get": {
                "tags": [
                    "Admin Category"
                ],
                "summary": "List categories",
                "description": "Get all categories",
                "operationId": "5b8af7aa48b4d04546a931aa7732910c",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/x-locale"
                    },
                    {
                        "$ref": "#/components/parameters/direction"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/limit"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Apply filter by category name.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "publish",
                        "in": "query",
                        "description": "Apply filter by publish status.",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all categories",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Category"
                ],
                "summary": "Create category",
                "description": "Create a new category",
                "operationId": "224b1e1d871b834875b093373d2e8657",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Engine Parts"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Spare parts related to engines"
                                    },
                                    "publish": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns created category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/admin/categories/{category}": {
            "get": {
                "tags": [
                    "Admin Category"
                ],
                "summary": "Get category details",
                "description": "Get one category by ID",
                "operationId": "0adbc1881edee733119128732f0b46da",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Category"
                ],
                "summary": "Update category",
                "description": "Update category by ID",
                "operationId": "84969c7990a1d367f39efa5fd546053d",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/x-locale"
                    },
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated category name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "publish": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns updated category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Category"
                ],
                "summary": "Delete category",
                "description": "Delete category by ID",
                "operationId": "2fc512c8dbad9e2d08262c2e50b82b6c",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns deleted category"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "description": "Category schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Granola"
                    },
                    "description": {
                        "type": "text",
                        "example": "its the best Granola"
                    },
                    "publish": {
                        "type": "boolean",
                        "example": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-12 15:20:32"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-14 15:21:32"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-15 15:22:32"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "direction": {
                "name": "direction",
                "in": "query",
                "description": "Specify sorting order.",
                "required": false,
                "schema": {
                    "type": "string",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                },
                "example": "asc"
            },
            "page": {
                "name": "page",
                "in": "query",
                "description": "Determine page number.",
                "required": false,
                "schema": {
                    "type": "integer"
                },
                "example": 1
            },
            "limit": {
                "name": "limit",
                "in": "query",
                "description": "Determine number of items per page.",
                "required": false,
                "schema": {
                    "type": "integer",
                    "enum": [
                        10,
                        20,
                        30,
                        40
                    ]
                },
                "example": 10
            },
            "x-locale": {
                "name": "x-locale",
                "in": "header",
                "required": false,
                "schema": {
                    "type": "string",
                    "enum": [
                        "en",
                        "ar"
                    ]
                },
                "example": "en"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "name": "Bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}